pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cerebras-cloud-sdk>=1.67.0",
    "faiss-cpu>=1.13.2",
    "gradio>=6.10.0",
    "numpy>=2.4.3",
    "sentence-transformers>=5.3.0",
]
